backend/pyproject.toml
[project]
name = "mongene-backend"
version = "0.1.0"
description = "MonGene - AI Math Problem Generation System Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "supabase>=2.11.0",
    "httpx>=0.28.0",
    "anthropic>=0.40.0",
    "openai>=1.60.0",
    "google-genai>=1.0.0",
    "matplotlib>=3.9.0",
    "numpy>=2.0.0",
    "reportlab>=4.2.0",
    "Pillow>=11.0.0",
    "python-multipart>=0.0.18",
    "sse-starlette>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
